tests/wpt/tests/html/canvas/offscreen/fill-and-stroke-styles/2d.gradient.object.compare.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.gradient.object.compare</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.gradient.object.compare</h1>
<p class="desc"></p>


<script>
var t = async_test("");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  var g1 = ctx.createLinearGradient(0, 0, 100, 0);
  var g2 = ctx.createLinearGradient(0, 0, 100, 0);
  _assertDifferent(g1, g2, "g1", "g2");
  ctx.fillStyle = g1;
  _assertSame(ctx.fillStyle, g1, "ctx.fillStyle", "g1");
  t.done();

});
</script>
